# TrackinExercise/src/main/swagger/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API used by the Trackin Exercise HMI",
    "version" : "v1",
    "title" : "Trackin Exercise API",
    "contact" : {
      "name" : "Quentin Supernant",
      "url" : "https://www.linkedin.com/in/quentin-supernant-a7008731",
      "email" : "qsupernant@gmail.com"
    }
  },
  "host" : "/",
  "basePath" : "/api",
  "tags" : [ {
    "name" : "waypoints"
  }, {
    "name" : "tours"
  }, {
    "name" : "drivers"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/api/drivers" : {
      "get" : {
        "tags" : [ "drivers" ],
        "summary" : "Return list of drivers sorted by lastName",
        "description" : "",
        "operationId" : "list",
        "produces" : [ "application/json; charset=UTF-8" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Driver"
            }
          },
          "412" : {
            "description" : "Not connected to database"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/api/tours" : {
      "get" : {
        "tags" : [ "tours" ],
        "summary" : "Return list of tours",
        "description" : "",
        "operationId" : "list",
        "produces" : [ "application/json; charset=UTF-8" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Tour"
            }
          },
          "412" : {
            "description" : "Not connected to database"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/api/tours/{tour_id}" : {
      "put" : {
        "tags" : [ "tours" ],
        "summary" : "Update a tour",
        "description" : "",
        "operationId" : "update",
        "parameters" : [ {
          "name" : "tour_id",
          "in" : "path",
          "description" : "tour identifier",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "tour json data",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "SUCCESS",
            "schema" : {
              "type" : "string"
            }
          },
          "412" : {
            "description" : "Not connected to database"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/api/tours/{tour_id}/waypoints" : {
      "get" : {
        "tags" : [ "tours" ],
        "summary" : "Return all waypoints from a tour",
        "description" : "",
        "operationId" : "listWaypoints",
        "produces" : [ "application/json; charset=UTF-8" ],
        "parameters" : [ {
          "name" : "tour_id",
          "in" : "path",
          "description" : "tour identifier",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Tour"
            }
          },
          "412" : {
            "description" : "Not connected to database"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/api/waypoints" : {
      "get" : {
        "tags" : [ "waypoints" ],
        "summary" : "Return list of waypoints sorted by position",
        "description" : "",
        "operationId" : "list",
        "produces" : [ "application/json; charset=UTF-8" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/WayPoint"
            }
          },
          "412" : {
            "description" : "Not connected to database"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      },
      "post" : {
        "tags" : [ "waypoints" ],
        "summary" : "Create a waypoint",
        "description" : "",
        "operationId" : "create",
        "produces" : [ "application/json; charset=UTF-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "waypoint json data",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "SUCCESS",
            "schema" : {
              "type" : "string"
            }
          },
          "412" : {
            "description" : "Not connected to database"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/api/waypoints/{waypoint_id}" : {
      "get" : {
        "tags" : [ "waypoints" ],
        "summary" : "Return a waypoint from id",
        "description" : "",
        "operationId" : "load",
        "produces" : [ "application/json; charset=UTF-8" ],
        "parameters" : [ {
          "name" : "waypoint_id",
          "in" : "path",
          "description" : "waypoint identifier",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/WayPoint"
            }
          },
          "412" : {
            "description" : "Not connected to database"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      },
      "put" : {
        "tags" : [ "waypoints" ],
        "summary" : "Update a waypoint",
        "description" : "",
        "operationId" : "update",
        "parameters" : [ {
          "name" : "waypoint_id",
          "in" : "path",
          "description" : "waypoint identifier",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "waypoint json data",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "SUCCESS",
            "schema" : {
              "type" : "string"
            }
          },
          "412" : {
            "description" : "Not connected to database"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      },
      "delete" : {
        "tags" : [ "waypoints" ],
        "summary" : "Remove a waypoint",
        "description" : "",
        "operationId" : "remove",
        "parameters" : [ {
          "name" : "waypoint_id",
          "in" : "path",
          "description" : "waypoint identifier",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "SUCCESS",
            "schema" : {
              "type" : "string"
            }
          },
          "412" : {
            "description" : "Not connected to database"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    }
  },
  "definitions" : {
    "Driver" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        },
        "gender" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Tour" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "driverId" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "WayPoint" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "label" : {
          "type" : "string"
        },
        "latitude" : {
          "type" : "string"
        },
        "longitude" : {
          "type" : "string"
        },
        "position" : {
          "type" : "integer",
          "format" : "int32"
        },
        "type" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    }
  }
}